requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.21.0
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anthropic==0.64.0
    # via langchain-anthropic
anyio==4.10.0
    # via
    #   anthropic
    #   groq
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via uvicorn
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via
    #   anthropic
    #   groq
    #   openai
fastapi==0.116.1
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
groq==0.31.0
    # via langchain-groq
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   anthropic
    #   groq
    #   langsmith
    #   openai
httpx-sse==0.4.1
    # via langchain-community
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jiter==0.10.0
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-anthropic==0.3.19
    # via -r requirements.in
langchain-community==0.3.27
    # via -r requirements.in
langchain-core==0.3.74
    # via
    #   langchain
    #   langchain-anthropic
    #   langchain-community
    #   langchain-groq
    #   langchain-openai
    #   langchain-text-splitters
langchain-groq==0.3.7
    # via -r requirements.in
langchain-openai==0.3.31
    # via -r requirements.in
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.15
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.2
    # via langchain-community
openai==1.101.0
    # via langchain-openai
orjson==3.11.2
    # via langsmith
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   groq
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via
    #   -r requirements.in
    #   langchain-community
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
regex==2025.7.34
    # via tiktoken
requests==2.32.5
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   groq
    #   openai
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
starlette==0.47.2
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
tiktoken==0.11.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   aiosqlite
    #   anthropic
    #   fastapi
    #   groq
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   -r requirements.in
    #   uvicorn
yarl==1.20.1
    # via aiohttp
zstandard==0.24.0
    # via langsmith
